requirements.txt
streamlit
openai
python-dotenv
pandas
langgraph
langchain-openai
pydantic
matplotlib
tabulate
azure-core
dbt-core
faiss-cpu
langchain
langchain-community
langchain-openai
plotly
sentence-transformers
numpy